refactoring user services
