+++ app/controllers/users.js
@@ -1,5 +1,5 @@
 const logger = require('../logger');
-const userService = require('../services/users');
+const userDB = require('../services/userDB');
 const { registerBodyMapper } = require('../mappers/users');
 const errors = require('../errors');
 const interactor = require('../interactors/users');
@@ -7,11 +7,11 @@ const interactor = require('../interactors/users');
 exports.registerUser = (req, res, next) => {
   logger.info('Starting the user creation');
   const user = registerBodyMapper(req.body);
-  return userService
+  return userDB
     .validateUser(user)
     .then(responseUser => {
       if (responseUser) return Promise.reject(errors.userExistsError('user already exists'));
-      return userService.registerUser(user).then(response => {
+      return userDB.registerUser(user).then(response => {
         logger.info(`User with name ${response.name}, and email ${response.name}, was created`);
         return res.status(201).send(response);
       });
